refactor: separate professor schedule display into its own page and file.
